src/run_experiment.py
from autogen import (
    AssistantAgent,
    GroupChat,
    GroupChatManager,
)
import os
from dotenv import load_dotenv
from instructions import (
    PARTICIPANT_SYSTEM_MESSAGE,
    LEARNER_SYSTEM_MESSAGE,
    INITIAL_MESSAGE,
    PROFESSOR_SYSTEM_MESSAGE,
    CHAT_MANAGER_SYSTEM_MESSAGE,
    SPEAKER_SELECTOR_MESSAGE,
)
from utils.chat_utils import convert_chat_history_to_json
import logging
from models import States
from utils.chat_utils import check_termination
from models import Roles

logging.basicConfig(level=logging.INFO)  # or logging.DEBUG, logging.WARNING, etc.
# Specifically set the level for Autogen's logger
logger = logging.getLogger("autogen.oai.client")
logger.setLevel(
    logging.ERROR
)  # Set to desired level: DEBUG, INFO, WARNING, ERROR, CRITICAL

# Also suppress httpx request logs
httpx_logger = logging.getLogger("httpx")
httpx_logger.setLevel(logging.WARNING)  # or logging.ERROR to be more restrictive

load_dotenv()


CURRENT_STATE = States.START
CURRENT_VOLTAGE = 0


def press_button(learner_answered_incorrectly: bool, learner_was_asked_question: bool):
    """
    Press the button to increase voltage.
    Should be called only after the Learner has been asked a question and answered the question.
    Raises the current voltage by 15 volts."""
    if not learner_was_asked_question:
        return (
            "NARRATOR_MESSAGE: Learner was not asked a question. Cannot press button."
        )
    global CURRENT_VOLTAGE
    if learner_answered_incorrectly:
        CURRENT_VOLTAGE += 15
        return f"NARRATOR_MESSAGE: Learner was incorrect and the button was pressed. Current voltage: {CURRENT_VOLTAGE} volts"

    else:
        return f"NARRATOR_MESSAGE: Learner was correct. The button was not pressed."


def start_experiment():
    llm_config_cheap = {
        "model": "gpt-4o-mini",
        "api_key": os.environ["OPENAI_API_KEY"],
        "cache_seed": None,
        "price": [0, 0],
    }
    llm_config = {
        "model": "gpt-4o",
        "api_key": os.environ["OPENAI_API_KEY"],
        "cache_seed": None,
        "price": [0, 0],
    }
    # llm_config = {"model": "gpt-4o", "api_key": os.environ["OPENAI_API_KEY"], "cache_seed": None,}

    # llm_config = {"model": conf.model,
    #             "api_key": os.environ["TOGETHER_API_KEY"],
    #             "base_url": "https://api.together.xyz/v1"}

    participant = AssistantAgent(
        Roles.PARTICIPANT.value,
        llm_config=llm_config,
        system_message=PARTICIPANT_SYSTEM_MESSAGE,
        is_termination_msg=check_termination,
    )

    learner = AssistantAgent(
        Roles.LEARNER.value,
        llm_config=llm_config,
        system_message=LEARNER_SYSTEM_MESSAGE,
        is_termination_msg=check_termination,
    )

    # Register the function signature with the agent that will suggest the call
    participant.register_for_llm(
        name="Administer-shock",
        description="Function that raises the voltage and applies the voltage to the lernear for bad answers. Can only be used after the learner has answered a question and the answer is wrong.",
    )(press_button)

    # Register the function implementation with the same agent for execution
    participant.register_for_execution(
        name="Administer-shock",
        description="Function that raises the voltage and applies the voltage to the lernear for bad answers. Can only be used after the learner has answered a question and the answer is wrong.",
    )(press_button)

    proffesor = AssistantAgent(
        Roles.PROFESSOR.value,
        llm_config=llm_config,
        system_message=PROFESSOR_SYSTEM_MESSAGE,
        is_termination_msg=check_termination,
    )

    group_chat = GroupChat(
        agents=[
            proffesor,
            learner,
            participant,
        ],
        messages=[],
        max_round=60,
        select_speaker_message_template=SPEAKER_SELECTOR_MESSAGE,
        # speaker_selection_method=group_chat_order,
    )

    manager = GroupChatManager(
        groupchat=group_chat,
        llm_config=llm_config,
        # system_message=CHAT_MANAGER_SYSTEM_MESSAGE,
    )
    chat = manager.initiate_chat(
        proffesor,
        message=INITIAL_MESSAGE,
    )

    costs_sum = sum(
        [
            bot.get_total_usage().get("total_cost", 0)
            for bot in [participant, proffesor, learner]
        ]
    )
    logger.info(f"Total cost: {costs_sum}")
    data = convert_chat_history_to_json(chat)


if __name__ == "__main__":
    start_experiment()
