pyproject.toml
[project]
name = "Milgram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.1.post0",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "openai>=1.82.1",
    "pillow>=11.2.1",
    "pyautogen>=0.9",
    "pygame>=2.6.1",
    "pygbag>=0.9.2",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.12",
    "uvicorn>=0.34.2",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
